requirements.txt
google-api-python-client
google-auth
google-generativeai
python-dotenv
pandas
openpyxl
